+++ Source/Models/Token.swift
@@ -6,6 +6,8 @@
 //  Copyright © 2017-2018 Omise Go Pte. Ltd. All rights reserved.
 //
 
+import BigInt
+
 /// Represents a token
 public struct Token {
 
@@ -17,7 +19,7 @@ public struct Token {
     public let name: String
     /// The multiplier representing the value of 1 token. i.e: if I want to give or receive
     /// 13 tokens and the subunitToUnit is 1000 then the amount will be 13*1000 = 13000
-    public let subUnitToUnit: Double
+    public let subUnitToUnit: BigInt
     /// Any additional metadata that need to be stored as a dictionary
     public let metadata: [String: Any]
     /// Any additional encrypted metadata that need to be stored as a dictionary
@@ -47,7 +49,7 @@ extension Token: Decodable {
         id = try container.decode(String.self, forKey: .id)
         symbol = try container.decode(String.self, forKey: .symbol)
         name = try container.decode(String.self, forKey: .name)
-        subUnitToUnit = try container.decode(Double.self, forKey: .subUnitToUnit)
+        subUnitToUnit = try container.decode(BigInt.self, forKey: .subUnitToUnit)
         metadata = try container.decode([String: Any].self, forKey: .metadata)
         encryptedMetadata = try container.decode([String: Any].self, forKey: .encryptedMetadata)
         createdAt = try container.decode(Date.self, forKey: .createdAt)
